add expr

+++ records/track_10min_16mb/2026-03-25_1100_13L_StructuredPruning/train_gpt.py
@@ -914,7 +914,7 @@ def measure_block_influence(
             start = s * seq_len
             if start + seq_len + 1 > val_tokens.numel():
                 break
-            input_ids = val_tokens[start:start + seq_len].unsqueeze(0).to(device)
+            input_ids = val_tokens[start:start + seq_len].unsqueeze(0).to(device).long()
 
             # Forward pass, collecting hidden states before/after each block
             x = model.tok_emb(input_ids)
